LPS-80145 Rename FragmentsEditorRender to FragmentEntryLinkList
