+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/css/fragments_editor/FragmentEntryLinkList.scss
@@ -1,6 +1,6 @@
 @import "./variables";
 
-.fragments-editor-render {
+.fragment-entry-link-list {
 	margin-top: $managementBarHeight;
 	padding: $fragmentEntryLinkControlHeight;
 }
