+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/css/main.scss
@@ -2,7 +2,7 @@
 
 @import "./fragments_editor/FragmentEntryLinkEditables";
 
-@import "./fragments_editor/FragmentsEditorRender";
+@import "./fragments_editor/FragmentEntryLinkList";
 
 @import "./fragments_editor/FragmentsEditorToolbar";
