+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/FragmentsEditor.soy
@@ -112,7 +112,7 @@
 		{/call}
 
 		<div class="contextual-sidebar-content">
-			{call com.liferay.layout.admin.web.FragmentsEditorRender.render}
+			{call com.liferay.layout.admin.web.FragmentEntryLinkList.render}
 				{param defaultEditorConfiguration: $defaultEditorConfiguration /}
 				{param defaultLanguageId: $defaultLanguageId /}
 				{param events: [
