+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/fragment_entry_link/FragmentEditableFieldTooltip.soy
