+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/fragment_entry_link/FragmentEntryLinkList.soy
@@ -1,7 +1,7 @@
-{namespace com.liferay.layout.admin.web.FragmentsEditorRender}
+{namespace com.liferay.layout.admin.web.FragmentEntryLinkList}
 
 /**
- * FragmentsEditorRender
+ * FragmentEntryLinkList
  */
 {template .render}
 	{@param defaultEditorConfiguration: ?}
@@ -35,7 +35,7 @@
 	]}
 
 	{let $entryLinksClasses kind="text"}
-		fragments-editor-render
+		fragment-entry-link-list
 
 		{if $translating}
 			{sp}translating
